requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
astor==0.8.1
async-timeout==4.0.3
attrs==24.2.0
av==12.3.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.7
contourpy==1.3.0
cryptography==43.0.1
ctranslate2==4.4.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
durationpy==0.9
emoji==2.8.0
exceptiongroup==1.2.2
fastapi==0.115.0
faster-whisper==1.0.3
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
groq==0.11.0
grpcio==1.66.2
h11==0.14.0
halo==0.0.31
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.4.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.8
langchain-experimental==0.3.2
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.130
lazy_loader==0.4
lmdb==1.5.1
log-symbols==0.0.14
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
openwakeword==0.6.0
opt-einsum==3.3.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
posthog==3.7.0
protobuf==3.20.2
pvporcupine==3.0.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2024.2
pywin32==306
PyYAML==6.0.2
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.1
rpds-py==0.20.0
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spinners==0.0.24
SQLAlchemy==2.0.35
stanza==1.6.1
starlette==0.38.6
stream2sentence==0.2.5
streamlit==1.39.0
streamlit-lottie==0.0.5
streamlit-on-Hover-tabs==1.0.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.2
torch==2.4.1
torchaudio==2.4.1
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
watchdog==5.0.3
watchfiles==0.24.0
webrtcvad==2.0.10
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
yarl==1.13.1
zipp==3.20.2
